Bound/requirements.txt
docx>=0.2.4
email-validator>=2.2.0
flask>=3.1.1
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
openai>=1.100.2
psycopg2-binary>=2.9.10
pypdf2>=3.0.1
python-docx>=1.2.0
sqlalchemy>=2.0.43
werkzeug>=3.1.3
python-dotenv>=1.0.0
